+++ packages/scrabble-solver/src/components/KeyMap/KeyMap.tsx
@@ -3,7 +3,7 @@ import { FunctionComponent } from 'react';
 import { selectConfig, useTranslate, useTypedSelector } from 'state';
 
 import Key from '../Key';
-import Sidebar from '../Sidebar';
+import Modal from '../Modal';
 
 import { Mapping } from './components';
 import { ARROWS, BACKSPACE, CTRL, DEL, ENTER, SPACE } from './keys';
@@ -19,8 +19,8 @@ const KeyMap: FunctionComponent<Props> = ({ className, isOpen, onClose }) => {
   const config = useTypedSelector(selectConfig);
 
   return (
-    <Sidebar className={className} isOpen={isOpen} title={translate('keyMap')} onClose={onClose}>
-      <Sidebar.Section title={translate('keyMap.board-and-rack')}>
+    <Modal className={className} isOpen={isOpen} title={translate('keyMap')} onClose={onClose}>
+      <Modal.Section title={translate('keyMap.board-and-rack')}>
         <Mapping description={translate('keyMap.board-and-rack.navigate')} mapping={[ARROWS]} />
         <Mapping description={translate('keyMap.board-and-rack.remove-tile')} mapping={[DEL, BACKSPACE]} />
         <Mapping description={translate('keyMap.board-and-rack.submit')} mapping={[ENTER]} />
@@ -39,17 +39,17 @@ const KeyMap: FunctionComponent<Props> = ({ className, isOpen, onClose }) => {
             ]}
           />
         )}
-      </Sidebar.Section>
+      </Modal.Section>
 
-      <Sidebar.Section title={translate('keyMap.board')}>
+      <Modal.Section title={translate('keyMap.board')}>
         <Mapping description={translate('keyMap.board.toggle-blank')} mapping={[SPACE, [CTRL, <Key key="b">B</Key>]]} />
         <Mapping description={translate('keyMap.board.toggle-direction')} mapping={[[CTRL, ARROWS]]} />
-      </Sidebar.Section>
+      </Modal.Section>
 
-      <Sidebar.Section title={translate('keyMap.rack')}>
+      <Modal.Section title={translate('keyMap.rack')}>
         <Mapping description={translate('keyMap.rack.insert-blank')} mapping={[SPACE]} />
-      </Sidebar.Section>
-    </Sidebar>
+      </Modal.Section>
+    </Modal>
   );
 };
