+++ packages/scrabble-solver/src/components/Modal/components/Section/Section.tsx
