+++ packages/scrabble-solver/src/components/Modal/components/Section/index.ts
