+++ packages/scrabble-solver/src/components/Modal/components/index.ts
