+++ packages/scrabble-solver/src/components/Settings/Settings.tsx
@@ -2,7 +2,7 @@ import { FunctionComponent } from 'react';
 
 import { useTranslate } from 'state';
 
-import Sidebar from '../Sidebar';
+import Modal from '../Modal';
 
 import { AutoGroupTilesSetting, ConfigSetting, LocaleSetting } from './components';
 
@@ -16,19 +16,19 @@ const Settings: FunctionComponent<Props> = ({ className, isOpen, onClose }) => {
   const translate = useTranslate();
 
   return (
-    <Sidebar className={className} isOpen={isOpen} title={translate('settings')} onClose={onClose}>
-      <Sidebar.Section title={translate('settings.game')}>
+    <Modal className={className} isOpen={isOpen} title={translate('settings')} onClose={onClose}>
+      <Modal.Section title={translate('settings.game')}>
         <ConfigSetting disabled={!isOpen} />
-      </Sidebar.Section>
+      </Modal.Section>
 
-      <Sidebar.Section title={translate('settings.language')}>
+      <Modal.Section title={translate('settings.language')}>
         <LocaleSetting disabled={!isOpen} />
-      </Sidebar.Section>
+      </Modal.Section>
 
-      <Sidebar.Section title={translate('settings.autoGroupTiles')}>
+      <Modal.Section title={translate('settings.autoGroupTiles')}>
         <AutoGroupTilesSetting disabled={!isOpen} />
-      </Sidebar.Section>
-    </Sidebar>
+      </Modal.Section>
+    </Modal>
   );
 };
