+++ packages/scrabble-solver/src/components/index.ts
@@ -10,6 +10,7 @@ export { default as KeyMap } from './KeyMap';
 export { default as Loading } from './Loading';
 export { default as Logo } from './Logo';
 export { default as Menu } from './Menu';
+export { default as Modal } from './Modal';
 export { default as NavButtons } from './NavButtons';
 export { default as NotFound } from './NotFound';
 export { default as PlainTiles } from './PlainTiles';
@@ -21,7 +22,6 @@ export { default as Results } from './Results';
 export { default as ResultsInput } from './ResultsInput';
 export { default as Screen } from './Screen';
 export { default as Settings } from './Settings';
-export { default as Sidebar } from './Sidebar';
 export { default as Solver } from './Solver';
 export { default as SolverMobile } from './SolverMobile';
 export { default as Splash } from './Splash';
