+++ packages/scrabble-solver/src/styles/variables.scss
@@ -77,7 +77,7 @@ $easeOutSine: cubic-bezier(0.61, 1, 0.88, 1);
   --focus--box-shadow--size: 6px;
 
   --dictionary--height: 260px;
-  --sidebar--width: 370px;
+  --modal--width: 370px;
   --solver-column--width: 580px;
   --square-button-size: 32px;
   --text-input--height: 40px;
